+++ apps/client/public/CHANGELOG.md
@@ -1,6 +1,12 @@
 Changelog is only provided in English language.
 If you found some bugs or want us to add new feature, please do so via [our GitHub Tracker](https://github.com/sakilabs/openbts/issues/new) or send me an email [ririxi@sakilabs.com](mailto:ririxi@sakilabs.com)
 
+# 2026-03-19
+
+### 🚀 Enhancements
+
+- **UKE source search** - searching on the map while UKE is selected as the source now queries permits and radiolines by permit ID or station ID
+
 # 2026-03-18
 
 ### 🚀 Enhancements